requirement.txt
accelerate>=1.2.1
peft>=0.14.0
bitsandbytes>=0.43.0
datasets>=3.2.0
deepspeed==0.15.4
einops>=0.8.0
liger_kernel==0.5.3
math-verify>=0.6.0
# lighteval @ git+https://github.com/huggingface/lighteval.git@657978d4e97830e94b5ad591918fbc2dfe0a635b
langdetect # Needed for LightEval's extended tasks
sentencepiece>=0.1.99
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
# --index-url https://download.pytorch.org/whl/cu118
transformers==4.49.0
trl==0.15.1
vllm==0.7.2
wandb>=0.19.1
notebook
pandas
unsloth
python-dotenv
colorlog
hydra-core>=1.3.2
tensorboard>=2.19.0
# flash-attn